Fixed bug regarding truncation of highest available address

+++ src/include/arch/memory/memoryMap.h
@@ -26,7 +26,7 @@ typedef void (*memoryMap_region_callback_t)(const memoryRegion_t *region);
 void memorymap_foreach_raw(memoryMap_region_callback_t callback);
 void memorymap_foreach_usable(memoryMap_region_callback_t callback);
 
-uintptr_t memorymap_get_highest_address(void);
+uint64_t memorymap_get_highest_address(void);
 
 void dump_memMap(void);
 void dump_usable_memory(void);
